# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vertex Finance OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountInput"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    put:
      operationId: updateAccount
      tags: [accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountInput"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CATEGORIES ──────────────────────────────────────────────────────────────
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List assets (investments)
      responses:
        "200":
          description: Assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    put:
      operationId: updateAsset
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /receivables:
    get:
      operationId: listReceivables
      tags: [assets]
      summary: List receivables
      responses:
        "200":
          description: Receivables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receivable"
    post:
      operationId: createReceivable
      tags: [assets]
      summary: Create receivable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReceivableInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receivable"
  /receivables/{id}:
    put:
      operationId: updateReceivable
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReceivableInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receivable"
    delete:
      operationId: deleteReceivable
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /debts:
    get:
      operationId: listDebts
      tags: [assets]
      summary: List debts
      responses:
        "200":
          description: Debts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debt"
    post:
      operationId: createDebt
      tags: [assets]
      summary: Create debt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDebtInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
  /debts/{id}:
    put:
      operationId: updateDebt
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDebtInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
    delete:
      operationId: deleteDebt
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /incomes:
    get:
      operationId: listIncomes
      tags: [assets]
      summary: List income sources
      responses:
        "200":
          description: Incomes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Income"
    post:
      operationId: createIncome
      tags: [assets]
      summary: Create income source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Income"
  /incomes/{id}:
    put:
      operationId: updateIncome
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncomeInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Income"
    delete:
      operationId: deleteIncome
      tags: [assets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── BUDGET ──────────────────────────────────────────────────────────────────
  /budget-groups:
    get:
      operationId: listBudgetGroups
      tags: [budget]
      summary: List budget groups
      responses:
        "200":
          description: Budget groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetGroup"
    post:
      operationId: createBudgetGroup
      tags: [budget]
      summary: Create budget group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetGroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetGroup"
  /budget-groups/{id}:
    put:
      operationId: updateBudgetGroup
      tags: [budget]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetGroupInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetGroup"
  /budget-items:
    get:
      operationId: listBudgetItems
      tags: [budget]
      summary: List budget items
      parameters:
        - name: groupId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Budget items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetItem"
    post:
      operationId: createBudgetItem
      tags: [budget]
      summary: Create budget item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetItem"
  /budget-items/{id}:
    put:
      operationId: updateBudgetItem
      tags: [budget]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetItem"

  # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [checking, savings, investment, wallet, other]
        balance:
          type: number
        color:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, balance, isActive, createdAt]
    CreateAccountInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [checking, savings, investment, wallet, other]
        balance:
          type: number
        color:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [name, type, balance]
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        category:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        status:
          type: string
          enum: [active, liquidated, pending]
        recurrence:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, description, category, amount, date, status, createdAt]
    CreateAssetInput:
      type: object
      properties:
        description:
          type: string
        category:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        status:
          type: string
          enum: [active, liquidated, pending]
        recurrence:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [description, category, amount, date, status]
    Receivable:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        category:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, received, overdue]
        recurrence:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, description, category, amount, dueDate, status, createdAt]
    CreateReceivableInput:
      type: object
      properties:
        description:
          type: string
        category:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, received, overdue]
        recurrence:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [description, category, amount, dueDate, status]
    Debt:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        creditor:
          type: string
        totalAmount:
          type: number
        remainingAmount:
          type: number
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, paid, renegotiated]
        monthlyInstallment:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, description, creditor, totalAmount, remainingAmount, dueDate, status, createdAt]
    CreateDebtInput:
      type: object
      properties:
        description:
          type: string
        creditor:
          type: string
        totalAmount:
          type: number
        remainingAmount:
          type: number
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, paid, renegotiated]
        monthlyInstallment:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required: [description, creditor, totalAmount, remainingAmount, dueDate, status]
    Income:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        source:
          type: string
        amount:
          type: number
        recurrence:
          type: string
          enum: [monthly, yearly, once, variable]
        isActive:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, description, source, amount, recurrence, isActive, createdAt]
    CreateIncomeInput:
      type: object
      properties:
        description:
          type: string
        source:
          type: string
        amount:
          type: number
        recurrence:
          type: string
          enum: [monthly, yearly, once, variable]
        isActive:
          type: boolean
        notes:
          type: string
          nullable: true
      required: [description, source, amount, recurrence]
    BudgetGroup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [fixed, variable, leisure, investment_lp, investment_cp, other]
        targetPercentage:
          type: number
          nullable: true
        color:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required: [id, name, type, sortOrder]
    CreateBudgetGroupInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [fixed, variable, leisure, investment_lp, investment_cp, other]
        targetPercentage:
          type: number
          nullable: true
        color:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required: [name, type, sortOrder]
    BudgetItem:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        groupName:
          type: string
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        description:
          type: string
        month:
          type: integer
        year:
          type: integer
        plannedAmount:
          type: number
        realizedAmount:
          type: number
        difference:
          type: number
        status:
          type: string
          enum: [on_track, over, under]
      required: [id, groupId, groupName, description, month, year, plannedAmount, realizedAmount, difference, status]
    CreateBudgetItemInput:
      type: object
      properties:
        groupId:
          type: integer
        categoryId:
          type: integer
          nullable: true
        description:
          type: string
        month:
          type: integer
        year:
          type: integer
        plannedAmount:
          type: number
        realizedAmount:
          type: number
      required: [groupId, description, month, year, plannedAmount, realizedAmount]
...