# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vertex Finance OS API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetStatus } from "./assetStatus";

export interface Asset {
  id: number;
  description: string;
  category: string;
  amount: number;
  date: Date;
  status: AssetStatus;
  recurrence?: string | null;
  notes?: string | null;
  createdAt: Date;
}