# lib/api-zod/src/generated/types/budgetItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vertex Finance OS API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetItemStatus } from "./budgetItemStatus";

export interface BudgetItem {
  id: number;
  groupId: number;
  groupName: string;
  categoryId?: number | null;
  categoryName?: string | null;
  description: string;
  month: number;
  year: number;
  plannedAmount: number;
  realizedAmount: number;
  difference: number;
  status: BudgetItemStatus;
}